Modals for Recipients CRUD finished

+++ EasySendler.WebUI/Controllers/RecipientsController.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
-using System.Linq;
-using System.Net;
-using System.Web;
+﻿using System.Linq;
 using System.Web.Mvc;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
@@ -36,21 +30,6 @@ namespace EasySendler.Controllers
             return View(db.Recipients.ToList());
         }
 
-        // GET: Recipients/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Recipient recipient = db.Recipients.Find(id);
-            if (recipient == null)
-            {
-                return HttpNotFound();
-            }
-            return View(recipient);
-        }
-
         [HttpGet]
         [JsonExceptionFilter]
         public JsonResult GetDetails(string id)
@@ -70,82 +49,21 @@ namespace EasySendler.Controllers
             };
         }
 
-        // GET: Recipients/Create
-        public ActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Recipients/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "RecipientId,Email,Name,SureName")] Recipient recipient)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Recipients.Add(recipient);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-
-            return View(Mapper.Map<RecipientsViewModel>(recipient));
-        }
-
-        // GET: Recipients/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Recipient recipient = db.Recipients.Find(id);
-            if (recipient == null)
-            {
-                return HttpNotFound();
-            }
-            return View(Mapper.Map<RecipientsViewModel>(recipient));
-        }
-
-        // POST: Recipients/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "rId,Email,Name,SureName")] RecipientsViewModel recipientsViewModel)
-        {
-            if (ModelState.IsValid)
-            {
-                Recipient recipient = db.Recipients.Find(recipientsViewModel.rId);
-                if (recipient != null)
-                {
-                    recipient.Email = recipientsViewModel.Email;
-                    recipient.Name = recipientsViewModel.Name;
-                    recipient.SureName = recipientsViewModel.SureName;
-
-                    db.SaveChanges();
-                }
-                
-                return RedirectToAction("Index");
-            }
-
-            return View(recipientsViewModel);
-        }
-
         [HttpGet]
         [JsonExceptionFilter]
         public ActionResult GetEdit(string id)
         {
-            int rlId;
-            if (!int.TryParse(id, out rlId))
+            int rId;
+            if (!int.TryParse(id, out rId))
             {
                 throw new JsonException("RecipientsController.GetDetails: id must be a number.");
             }
 
-            var result = db.Recipients.Where(x => x.RecipientId == rlId).AsQueryable().ProjectTo<RecipientsViewModel>().FirstOrDefault();
+            var result = rId == 0 
+                ? new RecipientsViewModel {rId = 0, Email = string.Empty, Name = string.Empty, SureName = string.Empty} 
+                : db.Recipients.Where(x => x.RecipientId == rId).AsQueryable().ProjectTo<RecipientsViewModel>().FirstOrDefault();
 
-            return PartialView(Mapper.Map<RecipientsViewModel>(result));
+            return PartialView(result);
         }
 
         [HttpPost]
@@ -154,33 +72,27 @@ namespace EasySendler.Controllers
         {
             if (ModelState.IsValid)
             {
-                Recipient recipient = db.Recipients.Find(recipientsViewModel.rId);
-                if (recipient != null)
+                if (recipientsViewModel.rId == 0)
                 {
-                    recipient.Email = recipientsViewModel.Email;
-                    recipient.Name = recipientsViewModel.Name;
-                    recipient.SureName = recipientsViewModel.SureName;
+                    db.Recipients.Add(Mapper.Map<Recipient>(recipientsViewModel));
 
                     db.SaveChanges();
+                }
+                else
+                {
+                    Recipient recipient = db.Recipients.Find(recipientsViewModel.rId);
+                    if (recipient != null)
+                    {
+                        recipient.Email = recipientsViewModel.Email;
+                        recipient.Name = recipientsViewModel.Name;
+                        recipient.SureName = recipientsViewModel.SureName;
+
+                        db.SaveChanges();
+                    }
                 }
             }
 
             return PartialView(recipientsViewModel);
-        }
-
-        // GET: Recipients/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Recipient recipient = db.Recipients.Find(id);
-            if (recipient == null)
-            {
-                return HttpNotFound();
-            }
-            return View(recipient);
         }
 
         // POST: Recipients/Delete/5
@@ -189,8 +101,13 @@ namespace EasySendler.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Recipient recipient = db.Recipients.Find(id);
-            db.Recipients.Remove(recipient);
-            db.SaveChanges();
+            if (recipient != null)
+            {
+                db.Recipients.Remove(recipient);
+
+                db.SaveChanges();
+            }
+            
             return RedirectToAction("Index");
         }
 
@@ -200,6 +117,7 @@ namespace EasySendler.Controllers
             {
                 db.Dispose();
             }
+
             base.Dispose(disposing);
         }
     }
